test: BoardTest에서 ViewTest를 분리
